Handle all instances of lost precision as gracefully as possible.
Types with increased precision are used where possible and in cases where adding increased precision wasn't practical, bounds checks were added. The behavior of the program should remain exactly the same unless side effects from implicit, lossy, type conversions were being relied upon. This is part of an effort to enable compilation with -Werror that will require several more patches, but this one is the largest.

+++ lib/plat/unix/unix-pipe.c
@@ -55,7 +55,7 @@ lws_plat_pipe_signal(struct lws_context *ctx, int tsi)
 	return eventfd_write(pt->dummy_pipe_fds[0], value);
 #else
 	char buf = 0;
-	int n;
+	ssize_t n;
 
 	n = write(pt->dummy_pipe_fds[1], &buf, 1);
